Rainbow.Framework.Data.MsSql added.
This is to become the main class that holds the dbml for the entire database. Eventually all the stored procs and other Linq or entity stuff will go in here for MsSql. Rainbow.Framework.Data will have a generic interface for this after we clean everything up to have all the data access go thru the MsSql stuff here.